+++ configuration.nix
@@ -268,7 +268,6 @@
     vim
     wget
     curl
-    aria2
     waybar
     pciutils
     gnome-keyring